requirements.txt
pillow
pyfirmata
pyclean
numpy
opencv-python
jupyter 